Agents/cover_letter_writer/requirements.txt
langchain_core==0.2.34
langchain_openai==0.1.22
python-dotenv==1.0.1
python_docx==1.1.2
